libs/narrator/requirements.txt
aniso8601==9.0.1
appdirs==1.4.3
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
CursesMenu==0.0.1a0
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
Flask==2.1.2
Flask-RESTful==0.3.9
gatorgrader==1.1.0
gitdb==4.0.9
GitPython==3.1.27
httplib2==0.14.0
hyperlink==19.0.0
ibm-cloud-sdk-core==3.15.1
ibmcloudant==0.1.3
idna==3.3
importlib-metadata==4.11.4
incremental==16.10.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
MarkupSafe==2.1.1
more-itertools==4.2.0
msgpack==1.0.4
netifaces==0.10.4
num2words==0.5.10
oauthlib==3.1.0
pbr==5.9.0
pexpect==4.6.0
pipenv==11.9.0
pipx==0.12.3.1
pkginfo==1.8.3
platformdirs==2.5.2
pluginbase==1.0.1
poetry-core==1.0.8
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.21.0
pycryptodome==3.15.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.4.0
pylev==1.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
shellingham==1.4.0
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.3
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
tomlkit==0.11.0
Twisted==18.9.0
typer==0.4.1
ubuntu-advantage-tools==27.8
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.1.2
zipp==1.0.0
zope.interface==4.7.1
